segmentacion: Agregadas entradas y pasado todo el procesamiento a floats. Agregado el blur dentro del algoritmo (difuminar con herramientas externas da errores de precisión).

+++ segmentation/disjoint-set.h
@@ -8,12 +8,12 @@ extern "C" {
 typedef void disjoint_set;
 
 disjoint_set* ds_new(size_t size);
-void ds_union(disjoint_set* ds, int a, int b, uint8_t diff);
+void ds_union(disjoint_set* ds, int a, int b, float diff);
 int ds_find(disjoint_set*, int a);
 
 /* NOTE: Este observador es específico del problema a resolver ¿Está bien
  *       ponerlo como parte de la biblioteca de disjoint_set? */
-int ds_idiff(disjoint_set*, int a, int K);
+float ds_idiff(disjoint_set*, int a, float K);
 
 }
