+++ segmentation/ds-arbol-compr.cpp
@@ -1,10 +1,11 @@
-#include<vector>
-#include<cstdint>
+#include <vector>
+#include <cstdint>
+#include <cmath>
 
 struct componente {
 	int padre;
 	int tamanio;
-	int idiff;
+	float idiff;
 };
 
 using disjoint_set = std::vector<struct componente>;
@@ -24,7 +25,7 @@ int ds_find(disjoint_set* ds, int a) {
 	return padre = ds_find(ds, padre);
 }
 
-void ds_union(disjoint_set* ds, int a, int b, uint8_t diff) {
+void ds_union(disjoint_set* ds, int a, int b, float diff) {
 	a = ds_find(ds, a);
 	b = ds_find(ds, b);
 
@@ -43,7 +44,7 @@ void ds_union(disjoint_set* ds, int a, int b, uint8_t diff) {
 	}
 }
 
-int ds_idiff(disjoint_set* ds, int a, int K) {
+float ds_idiff(disjoint_set* ds, int a, float K) {
 	a = ds_find(ds, a);
 	struct componente Ca = (*ds)[a];
 	return Ca.idiff + K / Ca.tamanio;
